refactor(phase-19)(pr-4a): combination 디렉터리 승격
crime_scene/combination.go (543줄) → crime_scene/combination/ 디렉터리 분할.

파일 구조:
- module.go — 구조체, NewCombinationModule, Init, Cleanup, engine.Module assertion, init()
- config.go/handlers.go/state.go/events.go (기능별 분리)
- testsupport.go — SeedCollectedForTest (테스트 seeding용 exported helper)

cross-package 통합 테스트 지원:
- crime_scene/integration_test.go에서 combo.collected (unexported) 직접 접근 제거
- combination.SeedCollectedForTest() 호출로 대체. evbus 구독과 동일한 효과.

test 이관:
- combination_test.go → combination/combination_test.go
- coverage_test.go의 combination 파트 (lines 417-559) → combination/coverage_test.go로 분리
- crime_scene/coverage_test.go 잔여 (location/evidence 테스트) 유지

Factory 서명, engine.Register 방식 불변.
F-go-3 해소 (500+ 파일 5→4).
