+++ apps/server/internal/module/crime_scene/combination/combination_test.go
@@ -1,4 +1,4 @@
-package crime_scene
+package combination
 
 import (
 	"context"
@@ -8,6 +8,14 @@ import (
 	"github.com/google/uuid"
 	"github.com/mmp-platform/server/internal/engine"
 )
+
+func newTestDeps() engine.ModuleDeps {
+	return engine.ModuleDeps{
+		SessionID: uuid.New(),
+		EventBus:  engine.NewEventBus(nil),
+		Logger:    nil,
+	}
+}
 
 func combinationCfg() json.RawMessage {
 	return json.RawMessage(`{
